feat: [Security] 요청에서 유저 객체 가져오기
