+++ spring-boot/src/main/java/com/tody/dayori/auth/controller/AuthController.java
@@ -6,6 +6,7 @@ import com.tody.dayori.config.JwtProvider;
 import com.tody.dayori.auth.dto.SignUpRequest;
 import com.tody.dayori.auth.service.UserService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -17,12 +18,14 @@ import java.net.URI;
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/auth")
+@Slf4j
 public class AuthController {
 
     private final UserService userService;
 
     @PostMapping("/login")
     public ResponseEntity<AuthResponse> login(@RequestBody LoginRequest loginRequest) {
+        log.debug("로그인 api 호출");
         return ResponseEntity.ok(userService.login(loginRequest));
     }
